Delete DummyJSONPost type
